pyproject.toml
[project]
name = "code-for-learning"
version = "0.1.0"
description = "LangGraph AI智能体开发实战 - 配套代码示例"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "fastmcp>=2.0.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "aiosqlite>=0.20.0",
    "matplotlib>=3.10.7",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.11",
    "langgraph-cli[inmem]>=0.4.8",
    "langchain-deepseek>=1.0.1",
    "tabulate>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["chapter2_agent_evaluation.py", "chapter2_cognitive_architecture.py"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"
